1.Change namespace qiankanglai.LoopScrollRect to Qiankanglai.LoopScrollRectManager, avoid conflict with LoopScrollRect.cs 2.Change BaseListBank to LoopListBankBase.cs, avoid conflict with Circular Scrolling List 3.Add InitOnStartBase that can Direct use, or override it 4.SizeHelper.cs from Demo/Scripts move to Scripts 5.LoopScrollRect.cs Add GetItemTypeStart() to get ItemTypeStart value
