+++ Assets/Demo/Scripts/UI_ControllData.cs
@@ -1,83 +1,113 @@
-﻿using SG;
-using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class UI_ControllData : MonoBehaviour
+namespace SG
 {
-    public Button m_ButtonAddData;
-    public Button m_ButtonSetData;
-    public Button m_ButtonDelData;
-    public Button m_ButtonSortData;
-    public Button m_ButtonReverseSortData;
-
-    public InitOnStart_Custom m_InitOnStart_Custom;
-
-    // Start is called before the first frame update
-    void Start()
+    public class UI_ControllData : MonoBehaviour
     {
-        m_ButtonAddData.onClick.AddListener(OnButtonAddDataClick);
-        m_ButtonSetData.onClick.AddListener(OnButtonSetDataClick);
-        m_ButtonDelData.onClick.AddListener(OnButtonDelDataClick);
-        m_ButtonSortData.onClick.AddListener(OnButtonSortDataClick);
-        m_ButtonReverseSortData.onClick.AddListener(OnButtonReverseSortDataClick);
-    }
+        public Button m_ButtonAddData;
+        public Button m_ButtonSetData;
+        public Button m_ButtonDelData;
+        public Button m_ButtonSortData;
+        public Button m_ButtonReverseSortData;
+        public Button m_ButtonSrollToCell;
 
-    private void OnDestroy()
-    {
-        m_ButtonAddData.onClick.RemoveListener(OnButtonAddDataClick);
-        m_ButtonSetData.onClick.RemoveListener(OnButtonSetDataClick);
-        m_ButtonDelData.onClick.RemoveListener(OnButtonDelDataClick);
-        m_ButtonSortData.onClick.RemoveListener(OnButtonSortDataClick);
-        m_ButtonReverseSortData.onClick.RemoveListener(OnButtonReverseSortDataClick);
-    }
+        public InputField m_InputFieldSrollToCell_CellIndex;
+        public InputField m_InputButtonSrollToCell_MoveSpeed;
 
-    private void OnButtonAddDataClick()
-    {
-        int RandomResult = Random.Range(0, 10);
-        m_InitOnStart_Custom.m_CustomListBank.AddContent(RandomResult);
+        public InitOnStart_Custom m_InitOnStart_Custom;
 
-        m_InitOnStart_Custom.m_LoopScrollRect.totalCount = m_InitOnStart_Custom.m_CustomListBank.GetListLength();
-        m_InitOnStart_Custom.m_LoopScrollRect.RefreshCells();
-    }
+        private CustomListBank m_CustomListBank;
 
-    private void OnButtonSetDataClick()
-    {
-        List<int> contents = new List<int>{
-            3, 4, 5, 6, 7
-        };
+        // Start is called before the first frame update
+        void Start()
+        {
+            m_CustomListBank = m_InitOnStart_Custom.m_ListBank as CustomListBank;
 
-        m_InitOnStart_Custom.m_CustomListBank.SetContents(contents);
-        m_InitOnStart_Custom.m_LoopScrollRect.totalCount = m_InitOnStart_Custom.m_CustomListBank.GetListLength();
-        m_InitOnStart_Custom.m_LoopScrollRect.RefillCells();
-    }
+            m_ButtonAddData.onClick.AddListener(OnButtonAddDataClick);
+            m_ButtonSetData.onClick.AddListener(OnButtonSetDataClick);
+            m_ButtonDelData.onClick.AddListener(OnButtonDelDataClick);
+            m_ButtonSortData.onClick.AddListener(OnButtonSortDataClick);
+            m_ButtonReverseSortData.onClick.AddListener(OnButtonReverseSortDataClick);
+            m_ButtonSrollToCell.onClick.AddListener(OnButtonSrollToCellClick);
+        }
 
-    private void OnButtonDelDataClick()
-    {
-        m_InitOnStart_Custom.m_CustomListBank.DelContentByIndex(0);
-        m_InitOnStart_Custom.m_LoopScrollRect.ClearCells();
-        m_InitOnStart_Custom.m_LoopScrollRect.totalCount = m_InitOnStart_Custom.m_CustomListBank.GetListLength();
-        m_InitOnStart_Custom.m_LoopScrollRect.RefillCells();
-    }
+        private void OnDestroy()
+        {
+            m_ButtonAddData.onClick.RemoveListener(OnButtonAddDataClick);
+            m_ButtonSetData.onClick.RemoveListener(OnButtonSetDataClick);
+            m_ButtonDelData.onClick.RemoveListener(OnButtonDelDataClick);
+            m_ButtonSortData.onClick.RemoveListener(OnButtonSortDataClick);
+            m_ButtonReverseSortData.onClick.RemoveListener(OnButtonReverseSortDataClick);
+            m_ButtonSrollToCell.onClick.RemoveListener(OnButtonSrollToCellClick);
+        }
 
-    private void OnButtonSortDataClick()
-    {
-        var TempContent = m_InitOnStart_Custom.m_CustomListBank.GetContents();
-        TempContent.Sort((x, y) => x.CompareTo(y));
+        private void OnButtonAddDataClick()
+        {
+            int RandomResult = Random.Range(0, 10);
+            m_CustomListBank.AddContent(RandomResult);
 
-        m_InitOnStart_Custom.m_LoopScrollRect.ClearCells();
-        m_InitOnStart_Custom.m_LoopScrollRect.totalCount = m_InitOnStart_Custom.m_CustomListBank.GetListLength();
-        m_InitOnStart_Custom.m_LoopScrollRect.RefillCells();
-    }
+            m_InitOnStart_Custom.m_LoopScrollRect.totalCount = m_InitOnStart_Custom.m_ListBank.GetListLength();
+            m_InitOnStart_Custom.m_LoopScrollRect.RefreshCells();
+        }
 
-    private void OnButtonReverseSortDataClick()
-    {
-        var TempContent = m_InitOnStart_Custom.m_CustomListBank.GetContents();
-        TempContent.Sort((x, y) => -x.CompareTo(y));
+        private void OnButtonSetDataClick()
+        {
+            List<int> contents = new List<int>
+            {
+                3, 4, 5, 6, 7
+            };
 
-        m_InitOnStart_Custom.m_LoopScrollRect.ClearCells();
-        m_InitOnStart_Custom.m_LoopScrollRect.totalCount = m_InitOnStart_Custom.m_CustomListBank.GetListLength();
-        m_InitOnStart_Custom.m_LoopScrollRect.RefillCells();
+            m_CustomListBank.SetContents(contents);
+            m_InitOnStart_Custom.m_LoopScrollRect.totalCount = m_InitOnStart_Custom.m_ListBank.GetListLength();
+            m_InitOnStart_Custom.m_LoopScrollRect.RefillCells();
+        }
+
+        private void OnButtonDelDataClick()
+        {
+            m_CustomListBank.DelContentByIndex(0);
+
+            int ShowLeftIndex = m_InitOnStart_Custom.m_LoopScrollRect.GetItemTypeStart();
+
+            m_InitOnStart_Custom.m_LoopScrollRect.ClearCells();
+            m_InitOnStart_Custom.m_LoopScrollRect.totalCount = m_InitOnStart_Custom.m_ListBank.GetListLength();
+            m_InitOnStart_Custom.m_LoopScrollRect.RefillCells();
+
+            m_InitOnStart_Custom.m_LoopScrollRect.SrollToCell(ShowLeftIndex, -1);
+            // No effect
+            //m_InitOnStart_Custom.m_LoopScrollRect.SrollToCell(ShowLeftIndex, 10);
+        }
+
+        private void OnButtonSortDataClick()
+        {
+            var TempContent = m_CustomListBank.GetContents();
+            TempContent.Sort((x, y) => x.CompareTo(y));
+
+            m_InitOnStart_Custom.m_LoopScrollRect.ClearCells();
+            m_InitOnStart_Custom.m_LoopScrollRect.totalCount = m_InitOnStart_Custom.m_ListBank.GetListLength();
+            m_InitOnStart_Custom.m_LoopScrollRect.RefillCells();
+        }
+
+        private void OnButtonReverseSortDataClick()
+        {
+            var TempContent = m_CustomListBank.GetContents();
+            TempContent.Sort((x, y) => -x.CompareTo(y));
+
+            m_InitOnStart_Custom.m_LoopScrollRect.ClearCells();
+            m_InitOnStart_Custom.m_LoopScrollRect.totalCount = m_InitOnStart_Custom.m_ListBank.GetListLength();
+            m_InitOnStart_Custom.m_LoopScrollRect.RefillCells();
+        }
+
+        private void OnButtonSrollToCellClick()
+        {
+            int Index = 0;
+            int.TryParse(m_InputFieldSrollToCell_CellIndex.text, out Index);
+
+            int MoveSpeed = 0;
+            int.TryParse(m_InputButtonSrollToCell_MoveSpeed.text, out MoveSpeed);
+
+            m_InitOnStart_Custom.m_LoopScrollRect.SrollToCell(Index, MoveSpeed);
+        }
     }
 }
