+++ Assets/Scripts/InitOnStartBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c48ab0d2f4a0b974eb3797139cf4dd68
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
